Moved screen capture into its own util class
